Add StethoInterceptor for OkHttp 2.2.0+
This makes it possible to enable Stetho network inspection by a one-line
change to existing projects that use OkHttp 2.2.0+.  This is made
possible by the new Interceptors feature:

  https://github.com/square/okhttp/wiki/Interceptors

Note that this diff also switches the sample app to use OkHttp for ease
of demonstration.  Eventually we should provide a stetho-hurlconnection
accessory module (akin to stetho-okhttp) which offers a more concise
wrapper for HttpURLConnection than our hand-coded demo.  Perhaps we
would also include stetho-httpclient although due to the deprecated
nature of this library (see
http://android-developers.blogspot.com/2011/09/androids-http-clients.html),
this is unlikely.