Avoid caching $HOME/.gradle lock files in Travis-CI
Without proper support for Gradle caching currently, it would seem the
best strategy for improving build times would be to remove the lock
files after build so that they are not detected as having changed.

Current build times are >2m and that seems really silly for Stetho.
Local builds on my machine are <50s.