support sending back rows changed
Previously we would only return a result for sql commands that were
selct statements. This diff makes `DatabasePeerManager` aware of what
commands are being executed and returns a result set appropriate to that
command.

Note: We are returning a result for all commands (including ones that do
not normally contain results like CREATE TABLE). We do this so that the
inspector interface will not remove the command.

Closes #82