Demonstrate an "optimal" release configuration with Stetho
This diff modifies stetho-sample and stetho-urlconnection to provide a
convenient configuration which excludes the stetho core module from
release builds but still allows the developer to ship network
instrumentation from stetho-urlconnection.

In particular, stetho-sample uses a separate Application instance for
debug builds only that initializes Stetho.  The trick is further
described here:
http://littlerobots.nl/blog/stetho-for-android-debug-builds-only/

For stetho-urlconnection, we allow it to be used unconditionally in
debug and release builds where only the debug builds actually include
stetho.  Stetho's presence is detected at runtime in and the internal
methods turn into no-ops accordingly.  This approach is intended to be
source and binary compatible with previous releases.  It is also
expected to efficiently pass the verifier on both Dalvik and ART without
spamming unresolved method/class warnings.  This is achieved by way of
the Holder class proxy in accordance with:
http://en.wikipedia.org/wiki/Initialization-on-demand_holder_idiom

Note that the strategy for stetho-okhttp is slightly different since the
hooking exists inside okhttp itself.  Because of this, callers can
arrange for a simple `src/debug/...` style solution to hook their
initialization of OkHttp much like how stetho-sample configures Stetho
itself.

Closes #16