Workaround serious throughput issue in Stetho's HTTP server
All communication between host and Stetho is done by a LocalSocket (a
UNIX abstract domain socket) which has a very heavy-handed
implementation of LocalSocketImpl#flush which polls/sleeps waiting for
the kernel buffer to become empty.  This behaviour in practice reduces
observed throughput of dumpapp (using either the buffered stdout or
stderr) to between 300 and 600KB/s despite adb being capable of ~5MB/s
on the same device.

This workaround involves blocking the final flush call to the socket
output stream.  This still drains the local Java buffer as originally
intended, but does not ask the kernel to drain its buffer (and block
doing so).  This kernel draining behaviour is entirely unnecessary for
our purposes and presumably for everyone else as it appears to be added
for some special purpose hack in the Bluetooh implementation.