Don't sleep forever in `dumpapp crash throw`
Wait on the crash thread instead, which will appropriately block the
command line until the crash is acknowledged by Android and takes down
the process.  Tested this with:

  former = Thread.getDefaultUncaughtExceptionHandler();
  Thread.setDefaultUncaughtExceptionHandler((...) -> {
    Thread.sleep(5000);
    former.uncaughtException(...);
  });

This caused `dumpapp crash throw` to hang for 5 seconds before the
process crashed and then the control was returned to the command line as
expected.

Separately, not invoking former.uncaughtException caused the process to
continue running, and `dumpapp crash throw` returns immediately (as
it should even if counter intuitively).