Implement Elements, Properties sub-tab
With this, developers can explore the Java object structure live from
Stetho as if a traditional debugger was attached.  This UI is a bit
buried and hard to notice, but can be found by clicking on an element in
the Elements tab and then finding the Properties tab on the right-hand
panel.  It is sometimes hidden behind a "more options"-style button.