Implement inspector button for the Elements tab
This allows the "inspector" magnifying glass to work on the elements tab. When
you click on the button, it enters a modal state where you can pick which UI
element should be highlighted. When you push your finger down on the Android
app, it will hit-test and highlight the appropriate view on the device (and
also as you move your finger around, if you'd like). When you move your finger
up, a command will be sent back to Chrome and it will now highlight the view
in the inspector UI.

Closes #124