Work around a subtle race condition in Database.getDatabaseTableNames
The protocol asks for a list of databases first then allows the user to
expand that at any later time to a list of tables.  The database may
have been removed however and there is seemingly no protocol support for
communicating this removal.  We have no choice but to yield a JSON-RPC
error.

Closes #154