Improve support pre-JB for Activity detection
This uses an older API (available to ICS+) to track activities instead
of relying on an ugly reflection trick known to work only on JB+ (and of
course not future proof).  There's a slight overhead increase in
Stetho.initialize as a result, but the timings added indicate it is
still very small (1-2ms on my Nexus 5).

The mechanism for tracking activities is partially exposed via
ActivityTracker which has a manual add/remove method which allows
developers to add support for pre-ICS even by hooking their own
application onCreate/onDestroy events just like HierarchyViewer's
ViewServer used in the past.

Closes #123