Introduce plugin framework for Console tab evaluation support
This allows third parties to provide their own REPL implementation to
interpret commands from the Console.  Eventually we'd like to explore
possibly using bsh or some other Java evaluation framework that is
enabled if detected in the classpath.  This is the first small step in
that direction.

Tested that exception processing and arbitrary object return types do
indeed work.