Fix crash when the support library is present but not used for a particular Activity.
This is done by simply checking whether we have an instance of
FragmentActivity before taking the support library path.  The framework
path doesn't need this check since it's impossible that if we have an
Activity and we are on a sufficiently high API level that the Activity
doesn't have fragment support.

Closes #176