Add allocation profiler to Stetho.
This introduces a UI that demonstrates how the functionality can be used
and adds support to do microscoped instrumentation with output to the
newly introduced Grimsey file format.

Ultimately the intention is to wire up a comprehensive toolchain that
can interoperate between Chrome's allocation profiler ("Heap Timeline")
and a more advanced / customized toolchain available for the Grimsey
format.