Split DOM into two classes: Document and DOM
DOM now handles only the communication between Chrome and Stetho, whereas Document is responsible for accessing, modifying, and maintaining information about elements on the Android side. The motivation for this change is that DOM is not the only class that needs access to information about elements - CSS also needs this information. By splitting these classes up it is possible for both DOM and CSS to be able to get and set element data.