Implement transparent request decompression
Request decompression is not officially part of the standard but many
mobile implementations choose to do it because payloads tend to be quite
large on clients and servers these days are easily made to support it.
This diff simply makes it so that Stetho will show the decompressed
payload, but still report the compressed payloads size to the UI.

Note that this only affects stetho-urlconnection and stetho-okhttp.
Other modules must manually implement this as it is not possible given
the existing API contract to do it transparently.

Closes #111