Further refine DefaultInspectorModulesBuilder
As the DevTools implementation has matured it has become clear that the
modules are not truly modular or interchangable at all and have heavy
dependencies between the two that make it impossible to provide or
configure them in isolation.  Because of this, we're moving to a more
explicit customization API which is responsible for managing the
dependencies of the implementation directly while still preserving the
level of abstractions already present in the system (such as providing a
custom document model, runtime REPL, etc).