Improved JSConsole when handling "%d" with a float as argument
In JavaScript every number is a floating point number, even ints! This causes
console.log("%d", 42) to fail because the initial implemenation was using
String.format() and it can't handle a floating point with %d.

We now have a custom formatter that mimics what google chrome does (with a few
improvemnts).

This patch also adds unit tests to the project stetho-js-rhino.