Remove dependency on Apache HttpClient
Android 6.0 removes support for HttpClient and forces applications to
bundle the jars directly if targeting 6.0 in an effort to phase out its
usage more aggressively.  Unfortunately, Stetho depended on HttpClient
for a crude HTTP server implementation to facilitate dumpapp and the
Chrome DevTools protocol.

This diff adds an HTTP server for general purpose usage within Stetho
but also heavily modifies dumpapp to no longer use HTTP at all.  This
became possible late in dumpapp's development when a framing protocol
was developed to support stdout/stderr.  This framing protocol made it
easy to use pure sockets and to further mimic the command-line paradigm.

The legacy HTTP support for dumpapp remains but in limited capacity
(does not support stdin and cannot do streaming stdout/stderr output).
A warning is produced when the old script is used.

Stetho combines both the new binary dumpapp protocol with the DevTools
HTTP/WebSocket protocol in a single socket and simply sniffs the first
few bytes from the client to determine how to handle the remainder of
the stream.  This was done for simplicity and performance so that just
one server would need to be set up when Stetho initializes.

Closes #251