No longer search for _dumpapp socket names
This is a cleanup diff to fix an issue accidentally left in
eb0ca3d565ce75f147d3ef8d376230f4e1f263a2.  Since we landed on just
having one socket, we don't need to support the extra *_dumpapp socket
name and in fact this script did so in a way that broke the -p use case.