Work around Travis memory restrictions with stetho-js-rhino
stetho-js-rhino is a huge artifact due to the external dependency on
Mozilla's Rhino package.  This causes extreme memory usage when dexing
the artifact as part of task
:stetho-js-rhino:transformClassesWithDexForDebugAndroidTest.  It is
believed that the cgroup policy applied to our job causes the process to
be killed as a result.

This change attempts to reduce the total resource load of the dx step to
sneak by this limit.

Fixes #354