Fix database id mapping to remove guesswork hacks
The SqliteDatabaseDriver had to resort to counter intuitive hacks
to manually locate a database by its file name and didn't
properly handle deduplicating multiple databases with the same name.  As
we expanded flexibility we failed to address the possibility (nay,
likelihood) that this will occur and it would create very confusing
results for users.

This diff doesn't address the confusion in the UI (two database entries
of the same name can still occur) but it internally makes it possible to
treat them as separate databases and track their filenames separately.
It also makes it relatively straightforward to fix the uniqueness of the
user visible naming by introducing a "context" parameter to
DatabaseDescriptor which would be used in the case of name ambiguity.

This diff addresses concerns raised in #366 and #367.