Have DefaultDatabaseProvider return filename
The API for `DatabaseProvider` is to return `List<File>` however the
current implementation of `DefaultDatabaseProvider` returns a `File`
that does not exist. We were able to get around this before due to the
fact that the file was only used to call `File#getName()`.

In #390 this was changed and the actual `File` object was used to open
the database.

This fixes #431