Add ability to highlight sub-bounds of view
A descriptor may not map directly to a View. For example if a
Descriptor was created for a Drawable it may occupy only a partial
space in the View. This diff solves this problem for descriptor
highlighting by extending HighlightableDescriptor with a method for
returning the bounds to highlight. Returning null will highlight the
full bounds of the View.