Add ability to register custom descriptors
This commit introduces the notion of a DescriptorProvider. A
DescriptorProvider is an object which given a DescriptorMap will
register a number of Descriptors. A list of DescriptorProviders can be
passed to the AndroidDocumentProviderFactory to add or override
descriptors for class of your choosing. This functionality enables
applications and libraries to add their own descriptors while relying
on Stetho defaults for all other settings.