Add ability for descriptors to customize element IDs
Currently a node id is directly connected to the instance of an object.
This has worked well up to this point. However we are implementing
custom descriptors which describe elements which are pooled under the
hood. When the UI changes we make no guarantees that a logical UI node
is represented by the same java object as before.

The problem with the current implementation in stetho is that this
causes the chrome element inspector to collapse/move nodes unexpectedly
when updating. This commit solve this by allows descriptors to
implement custom stable ids for nodes.