Move selector naming to the descriptor
Currently two selector names are hard-coded in DOM.java.
"<this_element>" and "Accessibility Properties". With this commit
descriptors are able to themselves choose what the selector name should
be and how their various properties should be distributed among the
selectors.

This enables an element to have styles such as:

layout {
  width: 100px;
  height: 100px;
}

props {
  text: "Hello World";
}

Due to this requiring an added property in StyleAccumulator I chose to
also take this commit to split up styles from computed styles which are
two separate concepts but up until now have been handled the same. A
style may be `width: 100%` while a computed style would be `width:
320px`.