Fix stdin issue when the stream is read to exhaustion.
This PR corrects a larger signedness mismatch between dumpapp client and
server.  The dumpapp client (python) used unsigned integers where the
server (Android/Java) used signed integers due to a limitation in
DataInput/OutputStream not supporting unsigned 32 and 64-bit types
(technically this could be done with readLong & 0xffffffffff though).
The critical fix comes in when sys.stdin.buffer.read reaches EOF whereby
the client tried to pack -1 in an unsigned int which is not obviously
not possible.  Since the server was always expecting this to be signed, simply
correcting the client causes the problem to go away.

Closes #509