Avoid false positive in LeakCanary
LeakCanary flags Stetho's ActivityTracker as a potential leak because it
does indeed have a path to the GC root that isn't already covered by the
many exclusions present in LeakCanary for legitimate Android use cases.
This is essentially a known flaw in LeakCanary in which unknown path to
GC roots are assumed to be leaks even if they are indeed consistent with
the normal object lifecycle (like in our case the strong reference is
removed when the activity is destroyed).

Working around this issue is trivial however and does indeed have the
effect of making our code look less suspicious of leaks in the first
place.

Closes #319