Add full "websocket" support to NetworkEventReporter
Note that websockets and sockets can be used more-or-less interchangably
from Stetho's point of view since we're emulating the support from
Chrome and therefore can choose to treat any socket-like stream of data
as a "websocket" for display purposes.  This is demonstrated in a
subsequent diff which introduces an IRC client to stetho-sample.

Note that there is no explicit pretty printing functionality as of yet
though this can be done at the client level by simply not hooking raw
sockets but rather interpreting the messages at a higher level before
sending them to NetworkEventReporter.