Add IRC client to stetho-sample to demonstrate new WebSocket support
The basic flow of the IRC client involves an initial connect screen
which prompts the user for server and nickname then transitions to a
chat screen which initiates the connection and offers a simple terminal
to type raw IRC commands and see the full set of replies from the
server.  Pressing back or encountering any I/O exception will cause the
user to return to the connect screen and will shutdown the websocket
(causing it to show as "Finished" in chrome://inspect).