Add analytics capabilities
We would like to track internal usage of stetho to figure out what to
focus on when improving our teams developer experience. This adds an
interface to do so.

I also added a single analytics log for when a method is called from
the inspector. This allows us to track which features are being used.